requirements.txt
joblib==1.4.2
numpy==2.0.2
pillow==11.0.0
scikit-learn==1.6.0
scipy==1.14.1
threadpoolctl==3.5.0
tqdm==4.67.1
